Create module JuliaConfig
